+++ vms/wizards/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2018, Jarsa Sistemas, S.A. de C.V.
+# -*- coding: utf-8 -*-
+# Copyright 2016, Jarsa Sistemas, S.A. de C
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 from . import vms_wizard_maintenance_order
-from . import vms_order_line_po
