--- vms/wizards/vms_order_line_po_view.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <record id="tms_wizard_invoice" model="ir.ui.view">
-        <field name="name">Create Purchase Order</field>
-        <field name="model">vms.order.line.po.wizard</field>
-        <field name="arch" type="xml">
-            <form string="Create Purchase Order">
-                <separator colspan="4" string="Create Purchase Order from External Lines ?"/>
-                <group>
-                    <field name="partner_id"/>
-                </group> 
-                <footer>
-                    <button name="make_po" string="Confirm" type="object"/>or
-                    <button special="cancel" string="Cancel"/>
-                </footer>
-            </form>
-        </field>
-    </record>
-    <act_window id="action_vms_order_po_wizard" key2="client_action_multi" name="Create Purchase Order" res_model="vms.order.line.po.wizard" src_model="vms.order" target="new" view_mode="form"/>
-</odoo>
